Remove bundled `liboping`, and link to system version instead.
Over the past several years, a vast majority of issues filed with this
crate have related to the build process with the bundled `liboping`;
usually this has to do with some autoconf-related issue. This is too
much of a support burden, and the crate authors (that's me) don't have
the resources to provide full support for all possible platforms. Thus,
this commit removes the bundled `liboping` and requires it to be
installed on the system first, with a (hopefully friendly!) error
message if it cannot be found.